P2 fix wide percentile grouping indexes
Finding: MEM-web-api-queries-009

Root cause:
Percentile and trimmed-mean grouping retain their series sizes and selected
slot counts as size_t, but both flush paths narrowed all traversal indexes to
int. A valid 64-bit grouping series above INT_MAX entries could therefore
produce truncated or negative indexes and read outside the series.

Classification and reachability:
This is a potential 64-bit memory-safety defect; no production occurrence was
demonstrated. The 86400 query limit caps output rows, not values accumulated
into one row, and single-row health, ML, badge, weights, API, and MCP queries
can cover the retained window. Successful allocation permits more than
INT_MAX doubles on 64-bit. On 32-bit, checked allocation arithmetic limits the
series below INT_MAX before the narrowing can be reached.

Solution:
Use ssize_t consistently for start, stop, last, interpolation, step, and loop
indexes and their count casts in both identical implementations. This signed
size domain represents every successfully allocated element and preserves the
intentional -1 stop sentinel used by reverse traversal.

Functional impact:
All values, flags, branches, interpolation decisions, reset behavior, growth,
and forward/reverse summation order remain identical for the previous valid
int range. Oversized valid series now stay in bounds. There is no API, ABI,
configuration, output, persistence, or failure-policy change.

Performance impact:
There are no added branches, operations, allocations, memory accesses, or loop
iterations. Complexity and the hot-path shape are unchanged; indexes use
native pointer width on 64-bit and remain int-width on 32-bit.

Caller, ownership, and concurrency impact:
time_grouping_flush(), rrd2rrdr_query_execute(), and rrd2rrdr() keep the same
return, flag, lifecycle, cancellation, and query-statistics contracts for all
health, ML, badge, weights, API v1/v2/v3, formatter, and MCP consumers.
Grouping state remains per-query ONEWAYALLOC memory, reset per row/dimension
and freed before return. No locks, atomics, shared state, ownership, or
lifetime behavior changes.

Validation:
- Native bit-comparison harness across empty, single, equal, positive,
  negative, mixed, forward/reverse, interpolation, and boundary cases
- 32-bit compile-only allocation/index representability check
- Affected production translation unit syntax compile with configured flags
- git diff and identical-pattern checks
- Six structured reviewer approvals with both blocking gates passing

(cherry picked from commit 25a780217965d52ea27f95f7bededa1feda4b790)

+++ src/web/api/queries/trimmed_mean/trimmed_mean.h
@@ -125,28 +125,28 @@ static inline NETDATA_DOUBLE tg_trimmed_mean_flush(RRDR *r, RRDR_VALUE_FLAGS *rr
                 percent_last_slot = 1 - percent_interpolation_slot;
             }
 
-            int start_slot, stop_slot, step, last_slot, interpolation_slot;
+            ssize_t start_slot, stop_slot, step, last_slot, interpolation_slot;
             if(min >= 0.0 && max >= 0.0) {
-                start_slot = (int)((available_slots - slots_to_use) / 2);
-                stop_slot = start_slot + (int)slots_to_use;
+                start_slot = (ssize_t)((available_slots - slots_to_use) / 2);
+                stop_slot = start_slot + (ssize_t)slots_to_use;
                 last_slot = stop_slot - 1;
                 interpolation_slot = stop_slot;
                 step = 1;
             }
             else {
-                start_slot = (int)available_slots - 1 - (int)((available_slots - slots_to_use) / 2);
-                stop_slot = start_slot - (int)slots_to_use;
+                start_slot = (ssize_t)available_slots - 1 - (ssize_t)((available_slots - slots_to_use) / 2);
+                stop_slot = start_slot - (ssize_t)slots_to_use;
                 last_slot = stop_slot + 1;
                 interpolation_slot = stop_slot;
                 step = -1;
             }
 
             value = 0.0;
-            for(int slot = start_slot; slot != stop_slot ; slot += step)
+            for(ssize_t slot = start_slot; slot != stop_slot ; slot += step)
                 value += g->series[slot];
 
             size_t counted = slots_to_use;
-            if(percent_interpolation_slot > 0.0 && interpolation_slot >= 0 && interpolation_slot < (int)available_slots) {
+            if(percent_interpolation_slot > 0.0 && interpolation_slot >= 0 && interpolation_slot < (ssize_t)available_slots) {
                 value += g->series[interpolation_slot] * percent_interpolation_slot;
                 value += g->series[last_slot] * percent_last_slot;
                 counted++;
